Add solar power querying
Solar power capacity has seen massive growth over the past few years, and it can
already occasionally provide a notable share of the total electricity produced.
Hence it is to be included from now on. Unfortunately, it is only currently
available as a forecasted value with hourly granularity, so that will have to do.